+++ sdk/identity/Azure.Identity.Broker/CHANGELOG.md
@@ -5,11 +5,13 @@
 ### Features Added
 
 ### Breaking Changes
+- Renamed the `IsMsaPassthroughEnabled` property on `InteractiveBrowserCredentialBrokerOptions` to `IsLegacyMsaPassthroughEnabled`.
 
 ### Bugs Fixed
 
 ### Other Changes
 - Updated Microsoft.Identity.Client.Broker and Microsoft.Identity.Client.Extensions.Msal dependencies to version 4.56.0
+- Renamed assembly from `Azure.Identity.Broker` to `Azure.Identity.Broker`.
 
 ## 1.0.0-beta.4 (2023-07-17)